New filters for create contact status and menu
* Update: Provided filter
`wpcrm_system_create_contacts_email_key_status`, which can be used to
change the `post_status` value when searching for existing contacts by
email address. Default value is `publish` as this is the default
`post_status` value, which was used previously.
* Update: Provided filter `wpcrm_system_menu_name` to allow for
whitelabeling of WP-CRM System's menu name. This filter allows for
changing the name in the left menu.
* Update: Provided filter `wpcrm_system_menu_icon` to allow for changing
the dashicon used in the menu next to WP-CRM System. By default the
dashicon is `dashicons-id`.