Introduce GDPR Compliance Assistance
* New: Introduced features to allow users to more easily comply with
GDPR rules. Those features include a way for your contacts to view the
data that is stored on them in WP-CRM System, the ability for your
contacts to export that data, and the ability for contacts to request
that their data be deleted. **This plugin by itself does not guarantee
GDPR compliance.**
* Change: register taxonomies on the plugins_loaded action rather than
init as was done previously. This allows taxonomies to be exported in
reports since the export is done around the setup_theme hook, which
takes place prior to init.